test as view based instead of model based
- moved test to NodeViewTests to account for the Pins being part of the DynamoViewModel
